close issue #159
close issue #401
resolves issue #390
resolves issue #176
resolves issue #388
improves issue #61

+++ EspionSpotify/Updater.cs
@@ -1,24 +1,23 @@
-﻿using EspionSpotify.Properties;
-using System;
-using System.Collections.Generic;
+﻿using System;
 using System.Configuration;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
+using EspionSpotify.Properties;
 
 namespace EspionSpotify
 {
-    internal class Updater
+    internal static class Updater
     {
-        internal const string UPDATER_DIRECTORY = "Updater";
-        internal const string UPDATER_TMP_DIRECTORY = "tmp_updater";
-        internal static string ProjectDirectory = AppDomain.CurrentDomain.BaseDirectory;
+        private const string UPDATER_DIRECTORY = "Updater";
+        private const string UPDATER_TMP_DIRECTORY = "tmp_updater";
+        private static readonly string ProjectDirectory = AppDomain.CurrentDomain.BaseDirectory;
 
-        internal static string UpdaterExtractedTempDirectoryPath = $"{ProjectDirectory}{UPDATER_TMP_DIRECTORY}";
-        internal static string UpdaterExtractedContentDirectoryPath = $@"{UpdaterExtractedTempDirectoryPath}\{UPDATER_DIRECTORY}";
-        internal static string UpdaterDirectoryPath = $"{ProjectDirectory}{UPDATER_DIRECTORY}";
+        private static readonly string UpdaterExtractedTempDirectoryPath = $"{ProjectDirectory}{UPDATER_TMP_DIRECTORY}";
+
+        private static readonly string UpdaterExtractedContentDirectoryPath =
+            $@"{UpdaterExtractedTempDirectoryPath}\{UPDATER_DIRECTORY}";
+
+        private static readonly string UpdaterDirectoryPath = $"{ProjectDirectory}{UPDATER_DIRECTORY}";
 
         internal static void UpgradeSettings()
         {
@@ -26,30 +25,33 @@ namespace EspionSpotify
             Settings.Default.Upgrade();
             Settings.Default.Save();
             UpdateDirectoryContent();
-            // DeleteOlderUserSettings();
+            DeleteOlderUserSettings();
         }
+
         private static void UpdateDirectoryContent()
         {
             try
             {
-                Directory.Delete(UpdaterDirectoryPath, recursive: true);
+                Directory.Delete(UpdaterDirectoryPath, true);
                 Directory.Move(UpdaterExtractedContentDirectoryPath, UpdaterDirectoryPath);
                 Directory.Delete(UpdaterExtractedTempDirectoryPath);
             }
-            catch { }
+            catch
+            {
+                // ignored
+            }
         }
 
         private static void DeleteOlderUserSettings()
         {
-            var path = Path.GetFullPath(Path.Combine(ConfigurationManager.OpenExeConfiguration(ConfigurationUserLevel.PerUserRoamingAndLocal).FilePath, @"..\..\"));
+            var path = Path.GetFullPath(Path.Combine(
+                ConfigurationManager.OpenExeConfiguration(ConfigurationUserLevel.PerUserRoamingAndLocal).FilePath,
+                @"..\..\"));
             var settingPaths = Directory.GetDirectories(path);
             foreach (var settingPath in settingPaths)
             {
-                DirectoryInfo info = new DirectoryInfo(settingPath);
-                if (info.Name != Application.ProductVersion)
-                {
-                    Directory.Delete(settingPath, true);
-                }
+                var info = new DirectoryInfo(settingPath);
+                if (info.Name != Application.ProductVersion) Directory.Delete(settingPath, true);
             }
         }
     }
